wip: fix compilation of 2.12
